% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinereg-package.R
\docType{package}
\name{vinereg-package}
\alias{vinereg-package}
\title{vinereg: D-vine regression models}
\description{
Fit D-vine copula regression models for conditional mean and quantile
prediction with continuous or discrete variables. The package provides
automatic covariate selection, fixed-order models, conditional distribution
diagnostics, and marginal fitted-quantile plots.
}
\references{
Kraus, D. and Czado, C. (2017). D-vine copula based quantile regression.
\emph{Computational Statistics & Data Analysis}, 110, 1--18.
\doi{10.1016/j.csda.2016.12.009}

Schallhorn, N., Kraus, D., Nagler, T., and Czado, C. (2017). D-vine
quantile regression with discrete variables. \doi{10.48550/arXiv.1705.08310}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tnagler.github.io/vinereg/}
  \item Report bugs at \url{https://github.com/tnagler/vinereg/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{mail@tnagler.com}

Authors:
\itemize{
  \item Thomas Nagler \email{mail@tnagler.com}
}

Other contributors:
\itemize{
  \item Dani Kraus [contributor]
}

}
