% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpit.R
\name{cpdf}
\alias{cpdf}
\title{Conditional PDF}
\usage{
cpdf(object, newdata, cores = 1)
}
\arguments{
\item{object}{an object of class \code{vinereg}.}

\item{newdata}{matrix of response and covariate values for which to compute
the conditional density}

\item{cores}{integer; the number of cores to use for computations.}
}
\description{
Calculates the conditional density of the response given the covariates.
}
\examples{
\dontshow{
set.seed(1)
}
# simulate data
x <- matrix(rnorm(100), 50, 2)
y <- x \%*\% c(1, -2)
dat <- data.frame(y = y, x = x, z = as.factor(rbinom(50, 2, 0.5)))

# fit vine regression model
fit <- vinereg(y ~ ., dat)

cpdf(fit, dat)
}
