% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.vinereg.R
\name{predict.vinereg}
\alias{predict.vinereg}
\title{Predict quantiles from a D-vine regression}
\usage{
\method{predict}{vinereg}(object, newdata, alpha = 0.5, uscale = F)
}
\arguments{
\item{object}{an object of class \code{vinereg}.}

\item{newdata}{matrix of covariate values for which to predict the quantile.}

\item{alpha}{vector of quantile levels.}

\item{uscale}{if \code{TRUE} input (newdata) and output is on copula scale.}
}
\value{
A vector of quantiles if alpha is a single number; a matrix of
quantiles with \code{length(alpha)} columns otherwise.
}
\description{
Predict quantiles from a D-vine regression
}

