% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.vinereg.R
\name{predict.vinereg}
\alias{predict.vinereg}
\title{Predict quantiles from a D-vine regression}
\usage{
\method{predict}{vinereg}(object, newdata, alpha = 0.5, uscale = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{vinereg}.}

\item{newdata}{matrix of covariate values for which to predict the quantile.}

\item{alpha}{vector of quantile levels.}

\item{uscale}{if \code{TRUE} input (newdata) and output is on copula scale.}

\item{...}{unused.}
}
\value{
A vector of quantiles if alpha is a single number; a matrix of
quantiles with \code{length(alpha)} columns otherwise.
}
\description{
Predict quantiles from a D-vine regression
}
\examples{
# simulate data
x <- matrix(rnorm(300), 100, 3)
y <- x \%*\% c(1, -1, 2)
dat <- data.frame(y = y, x = x, z = as.factor(rbinom(100, 3, 0.5)))

# fit vine regression model (parametric)
fit <- vinereg(y ~ ., dat, family_set = "parametric")

# model predictions (median)
pred <- predict(fit, newdata = dat, alpha = 0.5)

# observed vs predicted
plot(y, pred)

}
\seealso{
\code{\link{vinereg}}
}
