% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinereg.R
\name{vinereg}
\alias{vinereg}
\title{D-vine regressions}
\usage{
vinereg(y, x, familyset = "kde", correction = NA, par.1d = list(),
  cores = 1, ...)
}
\arguments{
\item{y}{response vector}

\item{x}{covariate matrix}

\item{familyset}{either \code{"kde"} for kernel estimation of the D-vine or
a vector of integers (see \code{\link{BiCopSelect}}).}

\item{correction}{correction criterion for the conditional log-likelihood.
\code{NA} (default) imposes no correction; other choices are \code{"AIC"}
and \code{"BIC"}.}

\item{cores}{integer.}

\item{...}{further arguments passed to \code{\link{kde1d}},
\code{\link{BiCopSelect}} or \code{\link{kdecop}}.}
}
\value{
An object of class \code{vinereg}.
}
\description{
D-vine regressions
}

