% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinereg.R
\name{vinereg}
\alias{vinereg}
\title{D-vine quantile regression}
\usage{
vinereg(formula, data, familyset = "kde", correction = NA,
  par_1d = list(), cores = 1, uscale = FALSE, ...)
}
\arguments{
\item{formula}{an object of class "formula"; same as [stats::lm()].}

\item{data}{matrix, data frame, list or environment (or object coercible by
[base::as.data.frame()]) containing the variables in the model.}

\item{familyset}{either \code{"kde"} for kernel estimation of the D-vine or a
vector of integers (see \code{\link{BiCopSelect}}).}

\item{correction}{correction criterion for the conditional log-likelihood.
\code{NA} (default) imposes no correction; other choices are \code{"AIC"}
and \code{"BIC"}.}

\item{par_1d}{list of options passed to [kdevine::kde1d()].}

\item{cores}{integer.}

\item{uscale}{logical indicating whether the data are already on copula scale
(no margins have to be fitted).}

\item{...}{further arguments passed to \code{\link{kde1d}},
\code{\link{BiCopSelect}} or \code{\link{kdecop}}.}
}
\value{
An object of class \code{vinereg}.
}
\description{
D-vine quantile regression
}
\examples{
# simulate data
x <- matrix(rnorm(300), 100, 3)
y <- x \%*\% c(1, -1, 2)
dat <- data.frame(y = y, x = x)

# fit vine regression model (parametric)
fit <- vinereg(y ~ ., dat, familyset = NA)

# model predictions (median)
pred <- predict(fit, newdata = dat, alpha = 0.5)

# observed vs predicted
plot(y, pred)

}
\seealso{
\code{\link{predict.vinereg}}
}
